[Adds] some of the crypto and a hidden text field.
Adds some of the crypto in a form that may or may not work or be kept,
mostly to see if the functions work. Also adds some http, which doesn’t
entirely work at the moment. Finally, adds a hidden text field to the
password form to stop it submitting on enter, which was using a GET
instead of a POST, and causing all sorts of annoying behavior.